% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_plot.R
\name{graph_plot}
\alias{graph_plot}
\title{Barplots}
\usage{
graph_plot(df, var_plot, var_label, var_grid = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{var_plot}{character. the variable for the x-axis}

\item{var_label}{character. the variable for the y-axis}

\item{var_grid}{character. the variable you will use to split the graph (if you do)}
}
\description{
Creates barplots for 2 or 3 variables
}
