% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_bins.R
\name{number_to_bins}
\alias{number_to_bins}
\title{Data wrangling}
\usage{
number_to_bins(
  df,
  cols = c("Age", "UnitLengthStay"),
  bins = list(c(0, 10, 20, 40, 60, 70, 80, 110), c(0, 1, 2, 3, 7, 15, 200))
)
}
\arguments{
\item{df}{a dataframe}

\item{cols}{a vector of strings with the name of the columns with the numerical data}

\item{bins}{a list of numerical vector containing the limits of the bins for each variable}
}
\description{
Transforms a numerical column from a dataset in a categorical column
}
