% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_survive.R
\name{predict_survive}
\alias{predict_survive}
\title{Classification Model}
\usage{
predict_survive(
  df,
  df_new_pacients = NULL,
  prop_train = 0.8,
  seed = 9238,
  features = c("Age_bin", "Gender", "UnitLengthStay_bin", "IsArterialHypertension"),
  col_weight = c("UnitDischargeCode", "Age_bin", "Age_bin", "UnitLengthStay_bin"),
  val_weight = c("D", "0_to_10", "10_to_20", "15_to_200"),
  acc_weight = c(0.25, 0.25, 0.25, 0.25)
)
}
\arguments{
\item{df}{a dataframe}

\item{df_new_pacients}{a dataframe with new data you want to predict the outcome}

\item{prop_train}{numerical between 0 and 1. The proportion of the original dataset you will use to train the model}

\item{seed}{integer. the seed you want to use to replicate the same results}

\item{features}{a vector of characters. the features you will use in your model}

\item{col_weight}{a vector of strings. The columns on which rows you will apply the case.weights of the random forest model}

\item{val_weight}{a vector of strings. The values of the rows that identifies where to apply the case.weights}

\item{acc_weight}{a numerical vector. The case.weights you will apply}
}
\description{
Trains, tests and plots predictions a random forest model
}
