% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_class.R
\name{treat_class}
\alias{treat_class}
\title{Data wrangling}
\usage{
treat_class(
  df,
  col_cat = NULL,
  old_cat = list(c("0", "1")),
  new_cat = list(c("NO", "YES"))
)
}
\arguments{
\item{df}{a dataframe}

\item{col_cat}{a vector of strings. The column you want to rename the factor (if you do)}

\item{old_cat}{A list of vectors of strings. The label of the factors you want to rename}

\item{new_cat}{A list of vectors of strings. The new label you want to create}
}
\description{
Fixes the classes of the columns of a dataset
}
