% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_na.R
\name{treat_na}
\alias{treat_na}
\title{Data wrangling}
\usage{
treat_na(
  df,
  view_only = F,
  plot_na = T,
  df_na = T,
  exclude_na = F,
  col_exclude = NULL,
  replace_na = F,
  col_replace = NULL,
  val_replace = NULL
)
}
\arguments{
\item{df}{a dataframe}

\item{view_only}{logical. If TRUE, doesn't change the dataframe}

\item{plot_na}{logical. If TRUE, plots the number of NA values by column}

\item{df_na}{logical. If TRUE, prints all the rows of the original dataframe that contains NA values}

\item{exclude_na}{logical. If TRUE, the rows wich cointains NA values are deleted}

\item{col_exclude}{character. The name of the column which contains the NA values you want to exclude}

\item{replace_na}{logical. If TRUE, the NA values will be replaced instead of deleted}

\item{col_replace}{character. The name of the column which contains the NA value you want to replace}

\item{val_replace}{A vector with the values you want to replace the NA observations(if you do)}
}
\description{
Prints, excludes or replaces NA values from a dataset
}
